Add "List of Requirements" tool using ReSpec hooks
This adds a "List of Requirements" as an additional navigation
section, after the table of contents.

This list is populated by scraping the specification document's
HTML markup.

Each list item is an occurrence of a "MUST" or "SHOULD" RFC 2119
keyword, along with some context.

List item headers link back to the respective section containing
the keyword.

The item elements themselves are clickable, and scroll to the
reference location in the document.

A "shortcut" button appears next to the ReSpec UI button, to allow
for jumping to the list. This also has a secondary purpose, as it
signals the existence of this feature.