Fix variable in recursive normalization
At 7.4.1 step 9.a.i, if `normalized` is a list containing `item` map, e.g.:

```json
[
    {
        "type": [
            "Person"
        ],
        "name": "Herman Melville"
    }
]
```

it seems that context for the recursive normalize data call should be `item["type"]`, not `keyValue["item"]`. In this case, context would be `Person`.

Same applies for step 9.b.i.

It appears `item["type"]` is being used in the reference implementation:

https://github.com/w3c/pub-manifest/blob/9a17073f65c5bae1e120933a9676250016b70148/experiments/manifest-to-internal-processor/manifestProcessor.js#L624